Develop

+++ bookingapi/src/main/java/com/poly/bookingapi/dto/VoucherDTO.java
@@ -18,7 +18,7 @@ public class VoucherDTO {
 
     private Integer status;
 
-    private BigDecimal requirement;
+    private Long requirement;
 
     private LocalDate startDate;
